+++ lib/inc/lib-utils-parse-number.h
@@ -12,7 +12,7 @@
 /*!
  * @brief Parse uint8_t from string.
  * @param str       String contains uint8_t to parse.
- * @param number    Pointer to on uint8_t to store value.
+ * @param number    Pointer on uint8_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_uint8(const char * str, uint8_t * number);
@@ -20,7 +20,7 @@ int parse_uint8(const char * str, uint8_t * number);
 /*!
  * @brief Parse int8_t from string.
  * @param str       String contains int8_t to parse.
- * @param number    Pointer to on int8_t to store value.
+ * @param number    Pointer on int8_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_int8(const char * str, int8_t * number);
@@ -28,7 +28,7 @@ int parse_int8(const char * str, int8_t * number);
 /*!
  * @brief Parse hexadecimal value on uint8_t from string.
  * @param str   String contains hexadecimal value on uint8_t to parse.
- * @param hex   Pointer to on uint8_t to store value.
+ * @param hex   Pointer on uint8_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_hex8(const char * str, uint8_t * hex);
@@ -36,7 +36,7 @@ int parse_hex8(const char * str, uint8_t * hex);
 /*!
  * @brief Parse uint16_t from string.
  * @param str       String contains uint16_t to parse.
- * @param number    Pointer to on uint16_t to store value.
+ * @param number    Pointer on uint16_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_uint16(const char * str, uint16_t * number);
@@ -44,7 +44,7 @@ int parse_uint16(const char * str, uint16_t * number);
 /*!
  * @brief Parse int16_t from string.
  * @param str       String contains int16_t to parse.
- * @param number    Pointer to on int16_t to store value.
+ * @param number    Pointer on int16_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_int16(const char * str, int16_t * number);
@@ -52,7 +52,7 @@ int parse_int16(const char * str, int16_t * number);
 /*!
  * @brief Parse hexadecimal value on uint16_t from string.
  * @param str   String contains hexadecimal value on uint16_t to parse.
- * @param hex   Pointer to on uint16_t to store value.
+ * @param hex   Pointer on uint16_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_hex16(const char * str, uint16_t * hex);
@@ -60,7 +60,7 @@ int parse_hex16(const char * str, uint16_t * hex);
 /*!
  * @brief Parse uint32_t from string.
  * @param str       String contains uint32_t to parse.
- * @param number    Pointer to on uint32_t to store value.
+ * @param number    Pointer on uint32_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_uint32(const char * str, uint32_t * number);
@@ -68,7 +68,7 @@ int parse_uint32(const char * str, uint32_t * number);
 /*!
  * @brief Parse int32_t from string.
  * @param str       String contains int32_t to parse.
- * @param number    Pointer to on int32_t to store value.
+ * @param number    Pointer on int32_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_int32(const char * str, int32_t * number);
@@ -76,7 +76,7 @@ int parse_int32(const char * str, int32_t * number);
 /*!
  * @brief Parse hexadecimal value on uint32_t from string.
  * @param str   String contains hexadecimal value on uint32_t to parse.
- * @param hex   Pointer to on uint32_t to store value.
+ * @param hex   Pointer on uint32_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_hex32(const char * str, uint32_t * hex);
@@ -84,7 +84,7 @@ int parse_hex32(const char * str, uint32_t * hex);
 /*!
  * @brief Parse uint64_t from string.
  * @param str       String contains uint64_t to parse.
- * @param number    Pointer to on uint64_t to store value.
+ * @param number    Pointer on uint64_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_uint64(const char * str, uint64_t * number);
@@ -92,7 +92,7 @@ int parse_uint64(const char * str, uint64_t * number);
 /*!
  * @brief Parse int64_t from string.
  * @param str       String contains int64_t to parse.
- * @param number    Pointer to on int64_t to store value.
+ * @param number    Pointer on int64_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_int64(const char * str, int64_t * number);
@@ -100,9 +100,17 @@ int parse_int64(const char * str, int64_t * number);
 /*!
  * @brief Parse hexadecimal value on uint64_t from string.
  * @param str   String contains hexadecimal value on uint64_t to parse.
- * @param hex   Pointer to on uint64_t to store value.
+ * @param hex   Pointer on uint64_t to store value.
  * @return 0 on success otherwise -1.
  */
 int parse_hex64(const char * str, uint64_t * hex);
+
+/*!
+ * @brief Parse double float value from string.
+ * @param str       String contains double float value to parse.
+ * @param number    Pointer on double to store value.
+ * @return 0 on success otherwise -1.
+ */
+int parse_double(const char * str, double * number);
 
 #endif /* LIB_UTILS_PARSE_NUMBER_H__ */