+++ CHANGELOG.md
@@ -7,7 +7,9 @@
  * Base library:
   * Assert,
   * String number parser,
+  * String utils,
   * Type define.
+ * Application skeleton
 
 ## Fixed